Minor refactoring on Rollback class
