+++ src/main/java/net/coreprotect/command/RollbackRestoreCommand.java
@@ -23,8 +23,8 @@ import net.coreprotect.config.Config;
 import net.coreprotect.config.ConfigHandler;
 import net.coreprotect.database.ContainerRollback;
 import net.coreprotect.database.Database;
-import net.coreprotect.database.Rollback;
 import net.coreprotect.database.lookup.PlayerLookup;
+import net.coreprotect.database.rollback.Rollback;
 import net.coreprotect.language.Phrase;
 import net.coreprotect.language.Selector;
 import net.coreprotect.utility.Chat;
